feat: switch to body.view (implement #285)
Switch from body.styled_view to body.view.
The change had a side effect on loosing the <div> with ID `Content`
which was aprovided by `body.styled_view` so we have to build it via
cheerio when the context is initiated.
Most of the test changes are triggered by this missing <div> on
the initial html.

Resolves: #285